requirements.txt
# Core dependencies
torch>=2.4.0
transformers>=4.45.0
datasets>=2.20.0
accelerate>=0.34.0

# Unsloth for efficient fine-tuning
unsloth>=2024.12

# LoRA and PEFT
peft>=0.13.0

# OpenAI API
openai>=1.50.0

# Training utilities
trl>=0.11.0
bitsandbytes>=0.44.0

# Evaluation
evaluate>=0.4.0
sqlparse>=0.5.0
sacrebleu>=2.4.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
loguru>=0.7.0

# API server (optional)
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0

# Experiment tracking (optional)
wandb>=0.18.0
